src/requirements.txt
Flask==1.1.2
flask-restplus==0.13.0
python-dotenv==0.17.0
Werkzeug==0.16.1
SQLAlchemy==1.3.12
Flask-Migrate==2.5.2